Other: Extract common UI and Editing parts as utils.
