+++ tests/fontfamily/fontfamilyediting.js
@@ -54,155 +54,6 @@ describe( 'FontFamilyEditing', () => {
 		} );
 	} );
 
-	describe( 'configuredOptions', () => {
-		it( 'should discard unparsable values', () => {
-			return VirtualTestEditor
-				.create( {
-					plugins: [ FontFamilyEditing ],
-					fontFamily: {
-						options: [ () => {}, 0, true ]
-					}
-				} )
-				.then( newEditor => {
-					editor = newEditor;
-
-					const plugin = editor.plugins.get( FontFamilyEditing );
-
-					expect( plugin.configuredOptions ).to.deep.equal( [] );
-				} );
-		} );
-
-		it( 'should pass through object definition', () => {
-			return VirtualTestEditor
-				.create( {
-					plugins: [ FontFamilyEditing ],
-					fontFamily: {
-						options: [
-							'default',
-							{
-								title: 'Comic Sans',
-								model: 'comic',
-								view: {
-									name: 'span',
-									style: {
-										'font-family': 'Comic Sans'
-									}
-								}
-							}
-						]
-					}
-				} )
-				.then( newEditor => {
-					editor = newEditor;
-
-					const plugin = editor.plugins.get( FontFamilyEditing );
-
-					expect( plugin.configuredOptions ).to.deep.equal( [
-						{
-							model: undefined,
-							title: 'Default'
-						},
-						{
-							title: 'Comic Sans',
-							model: 'comic',
-							view: {
-								name: 'span',
-								style: {
-									'font-family': 'Comic Sans'
-								}
-							}
-						}
-					] );
-				} );
-		} );
-
-		describe( 'shorthand presets', () => {
-			it( 'should return full preset from string presets', () => {
-				return VirtualTestEditor
-					.create( {
-						plugins: [ FontFamilyEditing ],
-						fontFamily: {
-							options: [
-								'Arial',
-								'"Comic Sans MS", sans-serif',
-								'Lucida Console, \'Courier New\', Courier, monospace'
-							]
-						}
-					} )
-					.then( newEditor => {
-						editor = newEditor;
-
-						const plugin = editor.plugins.get( FontFamilyEditing );
-
-						expect( plugin.configuredOptions ).to.deep.equal( [
-							{
-								title: 'Arial',
-								model: 'Arial',
-								view: {
-									name: 'span',
-									style: {
-										'font-family': 'Arial'
-									}
-								},
-								acceptsAlso: [
-									{
-										name: 'span',
-										style: {
-											'font-family': new RegExp( '("|\'|&qout;|\\W){0,2}Arial("|\'|&qout;|\\W){0,2}' )
-										}
-									}
-								]
-							},
-							{
-								title: 'Comic Sans MS',
-								model: 'Comic Sans MS',
-								view: {
-									name: 'span',
-									style: {
-										'font-family': '\'Comic Sans MS\', sans-serif'
-									}
-								},
-								acceptsAlso: [
-									{
-										name: 'span',
-										style: {
-											'font-family': new RegExp(
-												'("|\'|&qout;|\\W){0,2}Comic Sans MS("|\'|&qout;|\\W){0,2},' +
-												'("|\'|&qout;|\\W){0,2}sans-serif("|\'|&qout;|\\W){0,2}'
-											)
-										}
-									}
-								]
-							},
-							{
-								title: 'Lucida Console',
-								model: 'Lucida Console',
-								view: {
-									name: 'span',
-									style: {
-										'font-family': '\'Lucida Console\', \'Courier New\', Courier, monospace'
-									}
-								},
-								acceptsAlso: [
-									{
-										name: 'span',
-										style: {
-											'font-family': new RegExp(
-												'("|\'|&qout;|\\W){0,2}Lucida Console("|\'|&qout;|\\W){0,2},' +
-												'("|\'|&qout;|\\W){0,2}Courier New("|\'|&qout;|\\W){0,2},' +
-												'("|\'|&qout;|\\W){0,2}Courier("|\'|&qout;|\\W){0,2},' +
-												'("|\'|&qout;|\\W){0,2}monospace("|\'|&qout;|\\W){0,2}'
-											)
-										}
-									}
-								]
-							}
-						] );
-					} );
-			} );
-		} );
-	} );
-
 	describe( 'editing pipeline conversion', () => {
 		beforeEach( () => {
 			return VirtualTestEditor
